Use short[] for sound remapping

+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_13to1_12_2/packets/WorldPackets.java
@@ -339,7 +339,7 @@ public class WorldPackets {
     }
 
     public static int toNewId(int oldId) {
-        short newId = MappingData.blockMappings.getNewBlock(oldId);
+        int newId = MappingData.blockMappings.getNewBlock(oldId);
         if (newId != -1) {
             return newId;
         }
